% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rda_from_txt.R
\name{create_rda_from_txt}
\alias{create_rda_from_txt}
\title{This function read the txtfile first and create a rda object using the same name as txt file and save it into the data folder}
\usage{
create_rda_from_txt(txtfile, package_dir, rda_name)
}
\arguments{
\item{txtfile}{txtfile name and location}

\item{package_dir}{package directory}
}
\description{
This function read the txtfile first and create a rda object using the same name as txt file and save it into the data folder
}
\examples{
\dontrun{
create_rda_from_txt(txtfile='data_folder/sample.txt',package_dir = "mypackage/")
}
}

