% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESCRIPTION_make.R
\name{DESCRIPTION_make}
\alias{DESCRIPTION_make}
\title{This function create the DESCRIPTION file}
\usage{
DESCRIPTION_make(name_package, build_dir, title_package, version,
  author = "Yunlong", email = "nyunlong AT sfu.ca",
  description = "no discription")
}
\arguments{
\item{name_package}{package name}

\item{build_dir}{package directory}

\item{title_package}{packge title}

\item{version}{package version; a character if possible, like "1.1"}

\item{author}{author name}

\item{email}{email address}

\item{description}{description what the package does}
}
\description{
This function create the DESCRIPTION file
}
\examples{
\dontrun{
make_DESCRIPTION(name_package="mypackage",build_dir= "~/Dropbox/Rpackage/",title_package="My package",version="1.1")
}
}

