% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_push_win.R
\name{git_push_win}
\alias{git_push_win}
\title{This function automatically commit and push the changes on to github in window system}
\usage{
git_push_win(git_dir = "~/Dropbox/Yunlong/", commit = FALSE,
  comment = " ", keyword = "no_filter", limit_size = 5000,
  verbose = TRUE, type = c("deleted", "modified", "new file", "renamed"),
  git_exe = "git-cmd")
}
\arguments{
\item{git_dir}{repo directory}

\item{commit}{TRUE or FALSE, whether to commit or not}

\item{comment}{comment message, a string, only matters when commit is TRUE}

\item{keyword}{a string for keyword filtering, for you only want commit a certain folder each time}

\item{limit_size}{large file limit, default is 5000 which means 5MB}

\item{verbose}{TRUE or FALSE, print the git messages}

\item{type}{a string or a string vector, one of "deleted","modified","new file" and "renamed"}

\item{git_exe}{a string with the full name of git-cmd.exe, 'C:/Users/ynie01/AppData/Local/GitHub/PortableGit_c7e0cbde92ba565cb218a521411d0e854079a28c/git-cmd.exe'}
}
\description{
This function automatically commit and push the changes on to github in window system
}
\examples{
\dontrun{
git_push_win(git_dir=getwd(),commit=FALSE)
git_push_win(git_dir=getwd(),commit=TRUE,comment="first git push")
}
}

