% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_Rcode.R
\name{move_Rcode}
\alias{move_Rcode}
\title{This function moves the R source code into the R folder}
\usage{
move_Rcode(code_dir, name_package, build_dir, verbose = TRUE)
}
\arguments{
\item{code_dir}{R code dir, full name, put the all the R code here.}

\item{name_package}{package name}

\item{build_dir}{package directory}
}
\description{
This function moves the R source code into the R folder
}
\examples{
\dontrun{
move_Rcode(code_dir="Rcode/",name_package="my_packge",build_dir="~/Dropbox/Rpackage/")
}
}

